requirements.txt
ckanapi
python-dotenv
PyPDF2
openpyxl
geopandas
chardet